add edit_shop and update_shop actions to Admin::ProductCatgorizationsController
The behavior is now changed and it does not correspond to behavior
described in cukes. Instead of adding categories to shop one
by one we can now edit all alternative categories of given product
in one action.

I added a class dedicated for editing product's alternative
categories and it's specs.